requirements.txt
openai
pyreadstat
plotly
pandas
streamlit
matplotlib
python-dotenv